requirements.txt
# Core environment
python-dotenv
fastapi
uvicorn

# LangChain and vector store
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb

# Embeddings and LLM
openai
transformers
sentence-transformers

# PDF Loader
PyMuPDF

# Claude function calling (via your MCP server)
mcp 

# HTTP requests
requests


numpy==1.24.4


sentence-transformers